requirements.txt
telethon==1.24.0
python-dotenv==0.19.2
numpy==1.22.3
pandas==1.4.2
scikit-learn==1.0.2
transformers==4.18.0
torch==1.11.0
nltk==3.7
spacy==3.2.4
fastapi==0.75.1
uvicorn==0.17.6
python-telegram-bot==13.11
gunicorn==20.1.0
psycopg2-binary==2.9.3
redis==4.2.0
aiohttp==3.8.1
requests==2.27.1
beautifulsoup4==4.11.1
pymongo==4.1.1
sentence-transformers==2.2.0
joblib==1.1.0
dill==0.3.4
kagglehub==0.2.5
